Edit: Moved selectedGame contexts to view layer
