+++ src/modules/authentication/view/pages/Login/components/SignupForm/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
-// Components
+
 import { Input, SubmitButton } from 'shared/view/components';
 import Form from '../Form';
 
 // Libs
 import { Formik } from 'formik';
 
-// Hooks
+
 import useSignupController from 'modules/authentication/infra/controllers/useSignupController';
 
 // Schemas
@@ -15,7 +15,7 @@ import SignupSchema, {
   passwordsMatchValidation,
 } from 'modules/authentication/validation/Signup';
 
-// Styles
+
 import { signInTextStyle } from './styles';
 
 const initialValues = {
