+++ src/modules/authentication/view/pages/Login/index.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 
-// Components
+
 import { LoginForm, SignupForm } from './components';
 
-// Styles
+
 import {
   HomePage,
   TitleContainer,
