+++ src/modules/chooseGame/view/components/GameImage/index.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// Components
 import { Image, ImageStyle, StyleProp } from 'react-native';
 
 interface GameImageProps {
