+++ src/modules/chooseGame/view/pages/Lobby/components/EmptyList/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-// Styles
 import { Container, Text } from './styles';
 
 const EmptyList: React.FC = () => (
