+++ src/modules/chooseGame/view/pages/Lobby/components/Footer/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-// Hooks
+
 import { useNavigation } from '@react-navigation/native';
 import { useSessionContext } from 'shared/view/contexts';
 
-// styles
+
 import { Row, Button } from './styles';
 
 interface FooterProps {
