--- src/modules/selectedGame/container/contexts/PlayerProfileContext/index.ts
@@ -1,5 +1,0 @@
-import DefaultPlayerProfileContext from 'modules/selectedGame/container/contexts/PlayerProfileContext/implementations/DefaultPlayerProfileContext';
-
-const PlayerProfileContext = DefaultPlayerProfileContext;
-
-export default PlayerProfileContext;
