+++ src/modules/selectedGame/view/components/UserImage/index.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
-
-// Components
 import { Image, ImageStyle, StyleProp } from 'react-native';
 
 interface UserImageProps {
