+++ src/modules/selectedGame/view/contexts/PlayerProfileContext/index.ts
@@ -1,0 +1,4 @@
+import usePlayerProfileContext from 'modules/selectedGame/view/contexts/PlayerProfileContext/hooks/usePlayerProfileContext';
+import PlayerProfileContext from 'modules/selectedGame/view/contexts/PlayerProfileContext/implementations/DefaultPlayerProfileContext';
+
+export { PlayerProfileContext, usePlayerProfileContext };
