+++ src/modules/selectedGame/view/pages/PlayerProfile/components/AchievementCard/index.tsx
@@ -1,13 +1,9 @@
 import React from 'react';
 
-// Entities
 import { useNavigation } from '@react-navigation/native';
 
 import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
 
-// Hooks
-
-// Styles
 import { Container, Image, Text } from './styles';
 
 interface AchievementCardProps {
