+++ src/modules/selectedGame/view/pages/PlayerProfile/components/AchievementList/index.tsx
@@ -1,17 +1,14 @@
 import React, { useCallback, useMemo } from 'react';
-
-// Components
-import { Container, Title } from './styles';
 import { ActivityIndicator } from 'react-native';
-import AchievementCard from '../AchievementCard';
 
-// Helpers
-import { addObtainedFieldToAchievements } from './helpers';
-
-// Hooks
-import useGetAchievementsController from 'modules/selectedGame/infra/controllers/useGetAchievementsController';
-import usePlayerProfileContext from 'modules/selectedGame/container/contexts/PlayerProfileContext/contexts/usePlayerProfileContext';
 import { useFocusEffect } from '@react-navigation/core';
+
+import useGetAchievementsController from 'modules/selectedGame/infra/controllers/useGetAchievementsController';
+import { usePlayerProfileContext } from 'modules/selectedGame/view/contexts';
+
+import AchievementCard from '../AchievementCard';
+import { addObtainedFieldToAchievements } from './helpers';
+import { Container, Title } from './styles';
 
 const AchievementList: React.FC = () => {
   const { player } = usePlayerProfileContext();
