+++ src/modules/selectedGame/view/pages/PlayerProfile/components/Header/index.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
 
-// Hooks
-import usePlayerProfileContext from 'modules/selectedGame/container/contexts/PlayerProfileContext/contexts/usePlayerProfileContext';
+import { usePlayerProfileContext } from 'modules/selectedGame/view/contexts';
 import { useSessionContext } from 'shared/view/contexts';
 
-// Styles
 import { Container, Name, Title } from './styles';
 
 const Header: React.FC = () => {
   const { userData } = useSessionContext();
-  const player = usePlayerProfileContext();
+  const { player } = usePlayerProfileContext();
 
   return (
     <Container>
