+++ src/modules/selectedGame/view/pages/PlayerProfile/components/ProgressBar/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-// Styles
 import { BarContainer, Progress } from './styles';
 
 interface ProgressBarProps {
