+++ src/shared/view/components/Button/index.tsx
@@ -10,7 +10,6 @@ import {
 
 import { useThemeContext } from 'shared/view/contexts';
 
-// Styles
 import {
   RectContainer,
   StyledContainer,
